chore: Add WIP alert to page components
Add a WIP (Work in Progress) alert to the page components in the `[[id]]/+page.svelte` and `details/[id]/+page.svelte` files. This alert informs users that the site is being rebuilt with Turo as the backend. The alert includes a message and actions. This change improves the user experience by providing transparency about ongoing development work.